fix(web): rename cache module, fix config loading, expose valkey port
- Rename web/valkey.py → web/cache.py to avoid shadowing the valkey package
  (deferred import was a workaround; now import is top-level)
- Fix WebConfig.load() always used instead of from_env() so ./config.yaml
  is picked up when running uvicorn from NerdyPy/
- Expose valkey port 6379 in docker-compose for local dev
- Add docker.io registry prefix to image names for podman compatibility
- Add .run/WebApp.run.xml PyCharm run configuration
