+++ docker-compose.yml
@@ -59,8 +59,10 @@ services:
       - nerpybot-data:/data
 
   valkey:
-    image: valkey/valkey:8-alpine
+    image: docker.io/valkey/valkey:8-alpine
     restart: unless-stopped
+    ports:
+      - "6379:6379"
     volumes:
       - valkey-data:/data
     healthcheck:
@@ -105,7 +107,7 @@ services:
 
   # ── Uncomment to use PostgreSQL instead of SQLite ──
   # postgres:
-  #   image: postgres:18-alpine
+  #   image: docker.io/postgres:18-alpine
   #   restart: unless-stopped
   #   environment:
   #     POSTGRES_DB: nerpybot
